LPS-55505 move BaseScriptingExecutor to scripting-api bundle
