+++ modules/portal/portal-scripting-python/src/com/liferay/portal/scripting/python/internal/PythonExecutor.java
@@ -17,10 +17,10 @@ package com.liferay.portal.scripting.python.internal;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.SingleVMPool;
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayOutputStream;
-import com.liferay.portal.kernel.scripting.BaseScriptingExecutor;
 import com.liferay.portal.kernel.scripting.ExecutionException;
 import com.liferay.portal.kernel.scripting.ScriptingException;
 import com.liferay.portal.kernel.scripting.ScriptingExecutor;
+import com.liferay.portal.scripting.BaseScriptingExecutor;
 
 import java.util.HashMap;
 import java.util.Map;
